Reestructuración del proyecto
[BASE DEL NUEVO DISEÑO]
- Se eliminaro los componentes innecesarios
- Cambiar estilos
- El proyecto "PComplement" fue removido
- Router eliminado
- Paquetes actualizados
