+++ package.json
@@ -21,14 +21,14 @@
     "@angular/platform-browser": "~7.1.0",
     "@angular/platform-browser-dynamic": "~7.1.0",
     "@angular/router": "~7.1.0",
-    "angular-particle": "^1.0.4",
     "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
+    "ng2-animate-on-scroll": "^1.0.11",
+    "rxjs": "~6.5.1",
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.1",
+    "@angular-devkit/build-angular": "^0.13.8",
     "@angular/cli": "~7.1.4",
     "@angular/compiler-cli": "^7.2.8",
     "@angular/language-service": "~7.1.0",
