--- src/app/GLOBAL/navbar/navbar.component.html
@@ -1,36 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-dark bg-primary shadow-sm main-navbar py-0">
-  <a
-    class="navbar-brand py-0"
-    routerLink="">
-    <img src="assets/icon.png" width="60" alt="Azumi's Vault">
-  </a>
-  <button
-    class="navbar-toggler my-3"
-    type="button"
-    data-toggle="collapse"
-    aria-expanded="false"
-    aria-label="Toggle navigation"
-    (click)="isCollapsed = !isCollapsed">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-  <!-- Menú -->
-  <div class="collapse navbar-collapse" [ngbCollapse]="isCollapsed">
-    <ul class="navbar-nav ml-auto">
-      <a
-        class="nav-item nav-link py-2 py-md-3"
-        routerLink=""
-        routerLinkActive="active"
-        [routerLinkActiveOptions]="{exact: true}"
-        (click)="collapseNavbar()">
-        <i class="fas fa-home"></i> Inicio
-      </a>
-      <a
-        class="nav-item nav-link py-2 py-md-3"
-        routerLink="sobre-mi"
-        routerLinkActive="active"
-        (click)="collapseNavbar()">
-        <i class="fas fa-info-circle"></i> Sobre mí
-      </a>
-    </ul>
-  </div>
-</nav>
