--- src/app/about/about.component.html
@@ -1,45 +1,0 @@
-<particles [params]="myParams" [style]="myStyle" [width]="width" [height]="height"></particles>
-<div class="fade-in-top">
-  <div class="row my-1 my-md-4">
-    <div class="col-md-8">
-        <p class="title text-center text-md-left text-light">Cesar Alejandro Hernández Suárez</p>
-        <p class="text-center text-md-left text-about text-muted">Tecnólogo en Desarrollo de Software</p>
-      </div>
-      <div class="col-md-4">
-        <img class="img-fluid shadow rounded-circle d-block mx-auto" src="assets/images/profile.jpg" width="250">
-      </div>
-  </div>
-  <div class="bg-light row p-3 mt-3">
-    <div class="col-12">
-      <p class="title title-sm text-center text-md-left mb-0">Acerca de mí</p>
-      <p class="text-justify text-about">
-        ¡Muy buenas, mi nombre es <b>Alejandro</b> y te doy la bienvenida a mi portafolio de proyectos!<br>
-        Soy graduado del Centro de Enseñanza Técnica Industrial plantel Tonalá, o mejor conocido por sus siglas <b>CETI</b>,
-        lugar donde empecé a conocer este mundo tan maravilloso que es la programación.
-        Siempre me he visto fascinado por la tecnología y sus grandes avances al pasar los años, por lo cual, mi sueño desde pequeño ha sido
-        ser el creador/desarrollador de nuevas cosas que ayuden a la humanidad y mejoren nuestra <i>calidad de vida</i>.
-        Me encanta el desarrollo web, sobre todo en el apartado de <b>frontend</b>, ya que desde pequeño empezaba a tomar pequeñas plantillas web
-        y las modificaba a mi gusto, lamentablemente en ese tiempo no existía CSS3 ni HTML5, lo cual dependía mucho el uso de Adobe Flash
-        para contenidos dinámicos e interactivos.<br>
-        Comencé a programar desde los 15 años, poco después de que nos enseñaran pseudo-código en el CETI, comenzando por programación
-        lineal en C, más tarde aprendí nuevos paradigmas, como programación estructurada y orientada a objetos.<br>
-      </p>
-      <hr>
-      <p class="title title-sm text-center text-md-left">Información de contacto</p>
-      <!-- Enlaces -->
-      <div class="text-center">
-        <a
-        class="btn btn-outline-info web-link d-block d-lg-inline-block mr-0 mr-lg-2 mb-2 mb-lg-0"
-        href="mailto:alejandro-hdez115@outlook.com">
-          <i class="fas fa-envelope"></i> alejandro-hdez115@outlook.com
-        </a>
-        <a
-        class="btn btn-outline-info web-link d-block d-lg-inline-block"
-        href="https://github.com/AlexAzumi"
-        target="_blank">
-          <i class="fab fa-github"></i> AlexAzumi
-        </a>
-      </div>
-    </div>
-  </div>
-</div>
