--- src/app/app-routing.module.ts
@@ -1,34 +1,0 @@
-// Dependencias
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-// Componentes
-import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
-import { ProjectInfoComponent } from './project-info/project-info.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    pathMatch: 'full',
-    component: HomeComponent
-  },
-  {
-    path: 'sobre-mi',
-    component: AboutComponent
-  },
-  {
-    path: 'proyecto/:id',
-    component: ProjectInfoComponent
-  },
-  {
-    path: '**',
-    component: PageNotFoundComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
