+++ src/app/app.component.html
@@ -1,9 +1,9 @@
 <!-- Navbar -->
 <app-navbar></app-navbar>
 <!-- Router outlet -->
-<div
-  class="container-fluid pt-2 pb-4 pb-md-0">
-  <router-outlet></router-outlet>
+<div class="container pt-2 pb-4 pb-md-0">
+  <app-home></app-home>
+  <app-about></app-about>
 </div>
 <!-- Footer -->
-<app-footer></app-footer>
+<!-- <app-footer></app-footer> -->
