+++ src/app/app.module.ts
@@ -3,41 +3,30 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { HttpClientModule } from '@angular/common/http';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ParticlesModule } from 'angular-particle';
-// Router
-import { AppRoutingModule } from './app-routing.module';
 // Servicios
 import { ProjectsService } from './services/projects.service';
 // Componentes
 import { AppComponent } from './app.component';
-import { NavbarComponent } from './GLOBAL/navbar/navbar.component';
-import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { ProjectInfoComponent } from './project-info/project-info.component';
+import { NavbarComponent } from './components/GLOBAL/navbar/navbar.component';
+import { HomeComponent } from './components/home/home.component';
 // Pipes
 import { ReversePipe } from './pipes/reverse.pipe';
-import { FooterComponent } from './GLOBAL/footer/footer.component';
+import { FooterComponent } from './components/GLOBAL/footer/footer.component';
+import { AboutComponent } from './components/about/about.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     NavbarComponent,
     HomeComponent,
-    AboutComponent,
-    PageNotFoundComponent,
-    ProjectInfoComponent,
     ReversePipe,
-    FooterComponent
+    FooterComponent,
+    AboutComponent
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule,
     NgbModule,
-    HttpClientModule,
-    BrowserAnimationsModule,
-    ParticlesModule
+    HttpClientModule
   ],
   providers: [
     ProjectsService
