+++ src/app/components/GLOBAL/navbar/navbar.component.html
@@ -1,0 +1,41 @@
+<nav
+class="navbar navbar-expand-md navbar-dark bg-transparent main-navbar py-0 mt-md-4 pb-3 pb-md-4 border-bottom border-dark">
+  <button
+    class="navbar-toggler my-3"
+    type="button"
+    data-toggle="collapse"
+    aria-expanded="false"
+    aria-label="Toggle navigation"
+    (click)="isCollapsed = !isCollapsed">
+    <span class="navbar-toggler-icon"></span>
+  </button>
+  <!-- Menú -->
+  <div class="collapse navbar-collapse" [ngbCollapse]="isCollapsed">
+    <ul class="navbar-nav mx-auto">
+      <a
+        class="nav-item nav-link py-2 px-md-3 text-center rounded-pill mx-md-1 nav-fade-top-0"
+        href="#home"
+        (click)="collapseNavbar()">
+        Inicio
+      </a>
+      <a
+        class="nav-item nav-link py-2 px-md-3 text-center rounded-pill mx-md-1 nav-fade-top-1"
+        href="#about"
+        (click)="collapseNavbar()">
+        Sobre mí
+      </a>
+      <a
+        class="nav-item nav-link py-2 px-md-3 text-center rounded-pill mx-md-1 nav-fade-top-2"
+        href="#projects"
+        (click)="collapseNavbar()">
+        Proyectos
+      </a>
+      <a
+        class="nav-item nav-link py-2 px-md-3 text-center rounded-pill mx-md-1 nav-fade-top-3"
+        href="#contact"
+        (click)="collapseNavbar()">
+        Contacto
+      </a>
+    </ul>
+  </div>
+</nav>
