+++ src/app/components/about/about.component.html
@@ -1,0 +1,45 @@
+<div id="about">
+  <h3 class="d-inline-block about text-primary title mb-0 py-2 px-3 rounded-top">Sobre mí</h3>
+  <div class="row about rounded p-3 align-items-center">
+    <div class="col-md-8 text-light description">
+      <p class="mb-1">
+        Mi nombre es Alejandro Suárez, soy tecnólogo en desarrollo de software
+        egresado de CETI plantel Tonalá.<br>
+        Actualmente me he interesado mucho con el desarrollo web frontend, pero
+        también me fascinan otras tecnologías, como el desarrollo de videojuegos
+        y aplicaciones de escritorio.
+      </p>
+      <hr>
+      <p>Aquí hay un listado de las tecnologías y herramientas que he usado últimamente:</p>
+      <div class="row">
+        <div class="col">
+          <ul>
+            <li>Angular 6</li>
+            <li>Bootstrap 4</li>
+            <li>npm cli</li>
+            <li>SASS/SCSS</li>
+            <li>Netlify</li>
+          </ul>
+        </div>
+        <div class="col">
+          <ul>
+            <li>Unity</li>
+            <li>HTML5</li>
+            <li>CSS3</li>
+            <li>GitHub/GitLab</li>
+          </ul>
+        </div>
+      </div>
+      <hr class="mb-2">
+      <p class="text-info mb-0">
+        Tonalá, Jalisco
+      </p>
+    </div>
+    <div class="col-md-4">
+      <img
+        class="img-fluid rounded d-block mx-auto"
+        src="/assets/images/profile.jpg"
+        alt="Foto de Alejandro Suárez">
+    </div>
+  </div>
+</div>
