+++ src/app/components/home/home.component.html
@@ -1,0 +1,7 @@
+<header id="home" class="home my-large">
+  <div class="mb-3">
+    <h2 class="display-4 text-primary">Hola, mi nombre es</h2>
+    <h1 class="display-2 text-light font-weight-bold">Alejandro Suárez</h1>
+  </div>
+  <p class="h5 text-right text-info font-weight-normal">Tecnólogo en Desarrollo de Software</p>
+</header>
