--- src/app/home/home.component.html
@@ -1,20 +1,0 @@
-<particles [params]="myParams" [style]="myStyle" [width]="width" [height]="height"></particles>
-<p class="title text-center mb-0 pt-1 pb-3 pt-md-2 pb-md-4 text-light">Proyectos por <span class="text-info">Alejandro Suárez</span></p>
-<div class="row fade-in-top bg-light py-3" *ngIf="projects else noProjects">
-  <div class="col-md-6 my-2" *ngFor="let project of projects | reverse">
-    <div class="proyect-card border border-primary rounded bg-primary shadow-sm">
-      <a class="text-decoration-none" routerLink="proyecto/{{ project.id }}">
-        <img
-          class="img-fluid rounded-top"
-          src="assets/images/card-image/{{ project.id }}.jpg"
-          [alt]="project.title">
-        <p class="h2 text-center text-light pt-2 pt-md-1">{{ project.title }}</p>
-        <hr>
-        <p class="px-2 text-light">{{ project.description }}</p>
-      </a>
-    </div>
-  </div>
-</div>
-<ng-template #noProjects>
-  <p class="display-4 text-center text-light">No se han encontrado proyectos</p>
-</ng-template>
