--- src/app/page-not-found/page-not-found.component.html
@@ -1,10 +1,0 @@
-<particles [params]="myParams" [style]="myStyle" [width]="width" [height]="height"></particles>
-<div>
-  <div class="jumbotron shadow">
-    <h1 class="display-4">Página no encontrada</h1>
-    <p class="lead">La página solicitada no existe o fue removida del servidor</p>
-    <hr class="my-4">
-    <p>Para regresar al inicio puede dar clic en el siguiente botón</p>
-    <a class="btn btn-outline-primary btn-lg" routerLink="" role="button">Regresar al inicio</a>
-  </div>
-</div>
