--- src/app/project-info/project-info.component.html
@@ -1,51 +1,0 @@
-<div class="fade-in-top py-4" *ngIf="project else noProjectInfo">
-  <button
-    class="btn btn-outline-primary mb-3 mb-md-0"
-    routerLink="">
-    <i class="fas fa-arrow-left"></i> Regresar
-  </button>
-  <p class="title title-lg text-center mb-0">{{ project.title }}</p>
-  <p class="text-right project-date"><b>Fecha: </b>{{ project.date }}</p>
-  <hr>
-  <!-- Carousel -->
-  <ngb-carousel class="rounded" *ngIf="project.imagesURL">
-    <ng-template ngbSlide *ngFor="let image of project.imagesURL; index as i">
-      <img class="d-block w-100" [src]="image" alt="Imagen {{ i + 1 }}">
-    </ng-template>
-  </ngb-carousel>
-  <!-- Información -->
-  <div class="row border-top border-bottom mt-3">
-    <div class="col-lg-8 border-right">
-      <p class="title mb-0">Descripción del proyecto</p>
-      <p class="text-justify" [innerHTML]="project.description"></p>
-    </div>
-    <div class="col-md-4">
-      <p class="title title-sm mb-0">Tecnologías usadas</p>
-      <ul>
-        <li *ngFor="let tech of project.technologies">
-          {{ tech }}
-        </li>
-      </ul>
-    </div>
-  </div>
-  <!-- Enlaces -->
-  <div class="text-center pt-3">
-    <a
-    *ngIf="project.repository"
-    class="btn btn-outline-info web-link d-block d-md-inline-block mr-0 mr-md-4 mb-2 mb-md-0"
-    [href]="project.repository"
-    target="_blank">
-      <i class="fas fa-code-branch"></i> Ir al repositorio
-    </a>
-    <a
-      *ngIf="project.web"
-      class="btn btn-outline-info web-link d-block d-md-inline-block"
-      [href]="project.web"
-      target="_blank">
-        <i class="fas fa-globe-americas"></i> Ir a página web
-    </a>
-  </div>
-</div>
-<ng-template #noProjectInfo>
-  <p class="display-4 text-center">No se ha encontrado la información del proyecto</p>
-</ng-template>
