--- src/assets/data/project/pcomplement.json
@@ -1,17 +1,0 @@
-{
-    "id": "pcomplement",
-    "title": "PComplement",
-    "description": "Aplicación de escritorio, con base de datos MySQL y comunicada a través de ODBC el cual permite controlar el almacen de una empresa dedicada a la venta de componentes de computadora.",
-    "type": "Aplicación de escritorio (Windows)",
-    "date": "Junio de 2017",
-    "technologies": [
-        "C#",
-        "Base de datos MySQL",
-        "ODBC"
-    ],
-    "imagesURL": [
-        "assets/images/no-preview.jpg"
-    ],
-    "repository": null,
-    "web": null
-}
