+++ src/index.html
@@ -2,12 +2,12 @@
 <html lang="es">
 <head>
   <meta charset="utf-8">
-  <title>Azumi's Vault</title>
+  <title>Alejandro Suárez | Portafolio personal</title>
   <base href="/">
   <!-- Metadata -->
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="theme-color" content="#303F9F">
-  <meta name="description" content="Azumi's Vault - Portafolio de proyectos en línea de Alejandro Suárez">
+  <meta name="description" content="Alejandro Suárez - Portafolio de proyectos en línea">
   <link rel="manifest" href="/manifest.json">
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.7.2/css/all.css" integrity="sha384-fnmOCqbTlWIlj8LyTjo7mOUStjsKC4pOpQbqyi7RrhN7udi9RwhKkMHpvLbHG9Sr" crossorigin="anonymous">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
