+++ src/styles/custom.scss
@@ -1,134 +1,128 @@
 // Fuentes responsivas
-@import '../../node_modules/rfs/scss';
+@import "../../node_modules/rfs/scss";
 // Requerido
-@import '../../node_modules/bootstrap/scss/functions';
-@import '../../node_modules/bootstrap/scss/variables';
-@import '../../node_modules/bootstrap/scss/mixins';
+@import "../../node_modules/bootstrap/scss/functions";
+@import "../../node_modules/bootstrap/scss/variables";
+@import "../../node_modules/bootstrap/scss/mixins";
 // Fuentes
-@import url('https://fonts.googleapis.com/css?family=Asap|Assistant|Noto+Sans+TC|Ubuntu|Lato');
+@import url('https://fonts.googleapis.com/css?family=Poppins|Rubik|Nunito');
+
+// Opciones
+$enable-responsive-font-sizes: true;
 
 // Spacers
 $spacers: (
-    search: ($spacer * 6)
+  large: ($spacer * 12)
 );
 
 // Colores
-$primary: #303F9F;
+$primary: #03A9F4;
 
 // Tema
 $theme-colors: (
-    "primary": $primary
+  "primary": $primary
 );
 
-// General
+// Bootstrap
+@import "../../node_modules/bootstrap/scss/bootstrap";
+
+/*
+ * Global
+ */
+* {
+  font-family: 'Nunito', Arial, Helvetica, sans-serif;
+}
+
 html {
-    height: 100%;
+  height: 100%;
 }
 
 body {
-    min-height: 100%;
-    padding-bottom: 3.5rem;
-    position: relative;
+  background-color: $gray-800;
+  background-image: url('/assets/bg.png');
+  background-repeat: repeat;
+  background-attachment: fixed;
+  min-height: 100%;
+  padding-bottom: 3.5rem;
+  position: relative;
 }
 
-// Navbar
-.navbar-brand {
-    font-family: 'Lato', Arial;
+.title {
+  font-family: 'Rubik', Arial, Helvetica, sans-serif;
 }
 
+/*
+ * Navbar
+ */
 .nav-link {
-    font-family: 'Lato', sans-serif;
-    font-size: 1.2em;
-    transition: 150ms;
+  font-family: "Poppins", sans-serif;
+  font-size: 1.2em;
+  transition: 150ms;
 }
 
 .nav-link:hover {
-    background-color: $gray-900;
-    color: $white !important;
+  background-color: $gray-900;
+  color: $white !important;
 }
 
-// Cartas de proyectos
-.proyect-card {
-    transition: 150ms;
+// Animaciones
+.nav-fade-top-0 {
+  -webkit-animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 250ms both;
+  animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 250ms both;
 }
 
-.proyect-card:hover {
-    box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.5) !important;
+.nav-fade-top-1 {
+  -webkit-animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 500ms both;
+  animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 500ms both;
 }
 
-// Títulos
-.title {
-    @include rfs(40);
-    font-family: 'Ubuntu', sans-serif;
+.nav-fade-top-2 {
+  -webkit-animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 750ms both;
+  animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 750ms both;
 }
 
-.title-lg {
-    @include rfs(52);
+.nav-fade-top-3 {
+  -webkit-animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 1000ms both;
+  animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 1000ms both;
 }
 
-.title-sm {
-    @include rfs(30);
+/*
+ * Sobre mí
+ */
+.about {
+  background-color: rgba($color: $gray-800, $alpha: 0.8);
 }
 
-// Información de proyectos
-.project-date {
-    @include rfs(22);
-}
-
-.web-link {
-    @include rfs(30, true);
-}
-
-// Carousel
-.carousel {
-  border: 2px solid $gray-800;
-}
-
-// Footer
-.footer {
-    background-color: $primary;
-    bottom: 0;
-    color: $white;
-    left: 0;
-    position: absolute;
-    right: 0;
-}
-
-// Sobre mí
-.text-about {
-    font-size: 1.12em;
-}
-
-// Animaciones | Propiedad de Animista - http://animista.net
+/*
+ * Animaciones
+ * Propiedad de Animista - http://animista.net
+ */
 .fade-in-top {
-	-webkit-animation: fade-in-top 0.6s cubic-bezier(0.390, 0.575, 0.565, 1.000) 250ms both;
-	        animation: fade-in-top 0.6s cubic-bezier(0.390, 0.575, 0.565, 1.000) 250ms both;
+  -webkit-animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 250ms both;
+  animation: fade-in-top 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) 250ms both;
 }
 
 @-webkit-keyframes fade-in-top {
-    0% {
-      -webkit-transform: translateY(-20px);
-              transform: translateY(-20px);
-      opacity: 0;
-    }
-    100% {
-      -webkit-transform: translateY(0);
-              transform: translateY(0);
-      opacity: 1;
-    }
+  0% {
+    -webkit-transform: translateY(-20px);
+    transform: translateY(-20px);
+    opacity: 0;
   }
-  @keyframes fade-in-top {
-    0% {
-      -webkit-transform: translateY(-20px);
-              transform: translateY(-20px);
-      opacity: 0;
-    }
-    100% {
-      -webkit-transform: translateY(0);
-              transform: translateY(0);
-      opacity: 1;
-    }
+  100% {
+    -webkit-transform: translateY(0);
+    transform: translateY(0);
+    opacity: 1;
+  }
 }
-
-// Bootstrap
-@import '../../node_modules/bootstrap/scss/bootstrap';
+@keyframes fade-in-top {
+  0% {
+    -webkit-transform: translateY(-20px);
+    transform: translateY(-20px);
+    opacity: 0;
+  }
+  100% {
+    -webkit-transform: translateY(0);
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
